@@ -1,4 +1,8 @@
-<?php return array(
+<?php
+
+declare(strict_types=1);
+
+return [
 	0 => 'Абакум',
 	1 => 'Абрам',
 	2 => 'Абросим',
@@ -1389,17 +1393,17 @@
 	1388 => 'Рустем',
 	1389 => 'Карен',
 	1390 => 'Марсель',
-	1391 => "Юля",
-	1392 => "Мурат",
-	1393 => "Стас",
-	1394 => "Гамет",
-	1395 => "Галия",
-	1396 => "Асим",
-	1397 => "Гуля",
-	1398 => "Зафар",
-	1399 => "Даниэль",
-	1400 => "Саид",
-    1401 => "Маруся",
+	1391 => 'Юля',
+	1392 => 'Мурат',
+	1393 => 'Стас',
+	1394 => 'Гамет',
+	1395 => 'Галия',
+	1396 => 'Асим',
+	1397 => 'Гуля',
+	1398 => 'Зафар',
+	1399 => 'Даниэль',
+	1400 => 'Саид',
+    1401 => 'Маруся',
     1402 => 'Баир',
     1403 => 'Арзу',
-);
+];
