@@ -1,4 +1,8 @@
-<?php return array (
+<?php
+
+declare(strict_types=1);
+
+return [
   0 => 'Ааронов',
   1 => 'Абаимов',
   2 => 'Абакумов',
@@ -13653,5 +13657,5 @@
   13651 => 'Раднаев',
   13652 => 'Молнач',
   13653 => 'Гаар',
-  13654 => 'Марданов'
-);
+  13654 => 'Марданов',
+];
