@@ -1,4 +1,8 @@
-<?php return array (
+<?php
+
+declare(strict_types=1);
+
+return [
   0 => 'Абакумович',
   1 => 'Абакумовна',
   2 => 'Абрамович',
@@ -1555,4 +1559,4 @@
   1553 => 'Ярославовна',
   1554 => 'Ярославич',
   1555 => 'Ярославна',
-);
+];
